// src/Moryx.Resources.Web/app/src/app/api/models/resource-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* eslint-disable */
import { Entry, MethodEntry } from '@moryx/ngx-web-framework/entry-editor';

/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ResourceReferenceModel } from '../models/resource-reference-model';
export interface ResourceModel {
  description?: string | null;
  id?: number;
  methods?: Array<MethodEntry> | null;
  name?: string | null;
  partiallyLoaded?: boolean;
  properties?: Entry;
  referenceId?: number;
  references?: Array<ResourceReferenceModel> | null;
  type?: string | null;
}